+++ classes/search_perso.class.php
@@ -2,7 +2,7 @@
 // +-------------------------------------------------+
 // © 2002-2004 PMB Services / www.sigb.net pmb@sigb.net et contributeurs (voir www.sigb.net)
 // +-------------------------------------------------+
-// $Id: search_perso.class.php,v 1.23 2018-11-23 12:51:25 dgoron Exp $
+// $Id: search_perso.class.php,v 1.25.6.2 2019-10-10 08:05:17 dgoron Exp $
 
 if (stristr($_SERVER['REQUEST_URI'], ".class.php")) die("no access");
 
@@ -30,6 +30,9 @@ class search_perso {
 	public $search_perso_user;
 	public $directlink_user;
 	public $order;
+	protected $my_search;
+	protected $uses;
+	protected $error_message;
 	
 	// constructeur
 	public function __construct($id=0, $type='RECORDS') {
@@ -113,7 +116,8 @@ class search_perso {
 	}
 	
 	public function proceed() {
-		global $sub;
+		global $msg, $sub;
+		
 		switch($sub) {
 			case "form":
 				print $this->do_form();
@@ -134,8 +138,13 @@ class search_perso {
 				print $this->do_form();
 				break;
 			case "delete":
-				$this->delete();
-				print $this->do_list();
+				$deleted = $this->delete();
+				if($deleted) {
+				    print $this->do_list();
+				} else {
+				    print $this->get_display_header_list();
+				    error_message(	$msg[294], implode('<br />', $this->messages), 1, $this->get_url_base());
+				}
 				break;
 			case "launch":
 				// accès direct à une recherche personalisée
@@ -191,10 +200,10 @@ class search_perso {
 	}
 	
 	public function set_query() {
-		$my_search = $this->get_instance_search();
-		$this->query=$my_search->serialize_search();
-		$my_search->unserialize_search($this->query);
-		$this->human = $my_search->make_human_query();
+		$this->get_instance_search();
+		$this->query = $this->my_search->serialize_search();
+		$this->my_search->unserialize_search($this->query);
+		$this->human = $this->my_search->make_human_query();
 	}
 	
 	public function save() {
@@ -238,14 +247,14 @@ class search_perso {
 	// fonction générant le form de saisie 
 	public function do_form() {
 		global $msg,$tpl_search_perso_form,$charset;	
-		global $base_path, $current_module;
 		
 		// titre formulaire
 		if($this->id) {
 			$libelle=$msg["search_perso_form_edit"];
-			$link_duplicate="<input type='button' class='bouton' value='".$msg['duplicate']."' onClick=\"document.location='".$base_path."/".$current_module.".php?categ=search_perso&sub=duplicate&id=".$this->id."'\" />";
+			$link_duplicate="<input type='button' class='bouton' value='".$msg['duplicate']."' onClick=\"document.location='".$this->get_url_base()."&sub=duplicate&id=".$this->id."'\" />";
 			$link_delete="<input type='button' class='bouton' value='".$msg[63]."' onClick=\"confirm_delete();\" />";
 			$button_modif_requete = "";
+			$form_modif_requete = "";
 		} else {
 			$libelle=$msg["search_perso_form_add"];
 			$link_duplicate="";
@@ -253,9 +262,9 @@ class search_perso {
 			$button_modif_requete = "";
 			$form_modif_requete = "";
 			if(!$this->duplicate_from_id) {
-				$my_search = $this->get_instance_search();
-				$this->query=$my_search->serialize_search();
-				$this->human = $my_search->make_human_query();
+				$this->get_instance_search();
+				$this->query=$this->my_search->serialize_search();
+				$this->human = $this->my_search->make_human_query();
 			}
 		}
 		// Champ éditable
@@ -291,25 +300,41 @@ class search_perso {
 		
 		$link_annul = "onClick=\"unload_off();history.go(-1);\"";
 		$tpl_search_perso_form = str_replace('!!annul!!', $link_annul, $tpl_search_perso_form);
+		$tpl_search_perso_form = str_replace('!!url_base!!', $this->get_url_base(), $tpl_search_perso_form);
 		
 		return $tpl_search_perso_form;	
 	}
 
-	// fonction générant le form de saisie 
-	public function do_list() {
-		global $base_path;
-		global $msg;
-		global $action;
-		global $current_module;
-		
-		$display = "
+	protected function get_list_title() {
+	    global $msg;
+	    switch ($this->type) {
+	        case 'EXPL':
+	            return $msg["search_perso_expl_title"];
+	        default:
+	            return $msg["search_perso_title"];
+	    }
+	}
+	
+	public function get_display_header_list() {
+	    global $base_path;
+	    global $msg;
+	    
+	    $display = "
 		<script type='text/javascript' src='".$base_path."/javascript/search_perso_drop.js'></script>
-		<h1>".$msg["search_perso_title"]."</h1>
-		<div class='hmenu'>
-			<span><a href='./".$current_module.".php?categ=search_perso'>".$msg["search_perso_list_title"]."</a></span>".$this->directlink_user."
+		<h1>".$this->get_list_title()."</h1>
+        <div class='hmenu'>
+			<span><a href='".$this->get_url_base()."'>".$msg["search_perso_list_title"]."</a></span>".$this->directlink_user."
 		</div>
 		<hr />
 		<h3>".$msg["search_perso_list"]."</h3>";
+	    return $display;
+	}
+	
+	// fonction générant le form de saisie 
+	public function do_list() {
+		global $action;
+		
+		$display = $this->get_display_header_list();
 		switch ($action) {
 			case 'up':
 			case 'down':
@@ -331,17 +356,17 @@ class search_perso {
 		
 		if($this->type == 'AUTHORITIES') {
 			$searcher_tabs = new searcher_tabs();
-			$my_search=new search_authorities(true, 'search_fields_authorities');
+			$this->my_search=new search_authorities(true, 'search_fields_authorities');
 		} else {
-			$my_search=new search();
+			$this->my_search=new search();
 		}
 		$forms_search='';
 		$links='';
 		for($i=0;$i<count($this->search_perso_user);$i++) {
 			$target_url = $this->get_target_url($this->search_perso_user[$i]->id);
 			//composer le formulaire de la recherche
-			$my_search->unserialize_search($this->search_perso_user[$i]->query);
-			$forms_search.= $my_search->make_hidden_search_form($target_url,"search_form".$this->search_perso_user[$i]->id);
+			$this->my_search->unserialize_search($this->search_perso_user[$i]->query);
+			$forms_search.= $this->my_search->make_hidden_search_form($target_url,"search_form".$this->search_perso_user[$i]->id);
 			$libelle= $this->search_perso_user[$i]->name;
 			if($this->search_perso_user[$i]->directlink == 2) {
 				$js_launch_search= "document.forms['search_form".$this->search_perso_user[$i]->id."'].action += '&sub=launch';";
@@ -356,42 +381,106 @@ class search_perso {
 		return $forms_search.$links;
 	}
 
+	protected function get_uses_from_query($query) {
+	    $details = array();
+	    $result = pmb_mysql_query($query);
+	    if(pmb_mysql_num_rows($result)) {
+	        while($row = pmb_mysql_fetch_object($result)) {
+	            $search_query = unserialize($row->query);
+	            if(is_array($search_query)) {
+	                $criteria = $search_query['SEARCH'];
+	                if(is_array($criteria) && in_array('s_108', $criteria)) {
+	                    foreach ($criteria as $key=>$field) {
+	                        if($field == 's_1' && ($search_query[$key]['FIELD'][0] == $this->id)) {
+	                            $details[] = $row;
+	                        }
+	                    }
+	                }
+	            }
+	        }
+	    }
+	    return $details;
+	}
+	
+	protected function is_used() {
+	    global $msg;
+	    
+	    $this->uses = array();
+	    $this->messages = array();
+	    
+	    //DSI - Equations
+	    $query = "SELECT id_equation as id, nom_equation as label, requete as query FROM equations";
+	    $this->uses['equations'] = $this->get_uses_from_query($query);
+	    if(count($this->uses['equations'])) {
+	        $this->messages[] = $msg['search_perso_delete_used_by_equations'];
+	    }
+	    
+	    //Autre recherche prédéfinie
+	    $query = "SELECT search_id as id, search_name as label, search_query as query FROM search_perso WHERE search_id <> ".$this->id;
+	    $this->uses['search_perso'] = $this->get_uses_from_query($query);
+	    if(count($this->uses['search_perso'])) {
+	        $this->messages[] = $msg['search_perso_delete_used_by_search_perso'];
+	    }
+	    
+	    //Vues OPAC
+	    $query = "SELECT opac_view_id as id, opac_view_name as label, opac_view_query as query FROM opac_views";
+	    $this->uses['opac_views'] = $this->get_uses_from_query($query);
+	    if(count($this->uses['opac_views'])) {
+	        $this->messages[] = $msg['search_perso_delete_used_by_opac_views'];
+	    }
+	    
+	    if(count($this->uses['equations']) || count($this->uses['search_perso']) || count($this->uses['opac_views'])) {
+	        return true;
+	    }
+	    return false;
+	}
+	
 	// suppression d'une collection ou de toute les collections d'un périodique
 	public function delete() {
-		if($this->id) {
-			pmb_mysql_query("DELETE from search_perso WHERE search_id='".$this->id."' ");
-			$this->fetch_search_perso_user();
+	    if($this->id && !$this->is_used()) {
+	        pmb_mysql_query("DELETE from search_perso WHERE search_id='".$this->id."' ");
+	        $this->fetch_search_perso_user();
+	        return true;
 		}
+		return false;
 	}
 	
 	// fonction permettant d'accéder directement à une recherche prédéfinie
 	public function launch() {
 		if($this->id) {
-			$my_search=new search();
-			$my_search->unserialize_search($this->query);
-			print $my_search->make_hidden_search_form("./catalog.php?categ=search&mode=6","search_form".$this->id);
+			$this->my_search=new search();
+			$this->my_search->unserialize_search($this->query);
+			print $this->my_search->make_hidden_search_form("./catalog.php?categ=search&mode=6","search_form".$this->id);
 			print "<script type='text/javascript'>document.forms['search_form".$this->id."'].submit();</script>";
 		} else {
 			print $this->do_list();
 		}
 	}
 	
-	protected function get_instance_search() {
+	public function get_instance_search() {
+	    if (!empty($this->my_search)) {
+	        return $this->my_search;
+	    }
 		switch ($this->type) {
 			case 'AUTHORITIES':
-				$my_search=new search_authorities(true, 'search_fields_authorities');
+				$this->my_search=new search_authorities(true, 'search_fields_authorities');
 				break;
 			case 'EMPR':
-				$my_search=new search(true, 'search_fields_empr');
+				$this->my_search=new search(true, 'search_fields_empr');
 				break;
+			case 'EXPL':
+			    $this->my_search=new search(true, 'search_fields_expl');
+			    break;
 			default:
-				$my_search=new search();
+				$this->my_search=new search();
 				break;
 		}
-		return $my_search;
+		return $this->my_search;
 	}
 	
 	protected function get_target_url($id_predefined_search=0) {
+	    global $option_show_notice_fille, $option_show_expl;
+	    
 		switch ($this->type) {
 			case 'AUTHORITIES':
 				$searcher_tabs = new searcher_tabs();
@@ -400,6 +489,9 @@ class search_perso {
 			case 'EMPR':
 				$target_url = "./circ.php?categ=search";
 				break;
+			case 'EXPL':
+			    $target_url = "./catalog.php?categ=search&mode=8&option_show_notice_fille=$option_show_notice_fille&option_show_expl=$option_show_expl";
+			    break;
 			default:
 				$target_url = "./catalog.php?categ=search&mode=6";
 				break;
@@ -408,6 +500,11 @@ class search_perso {
 			$target_url .= "&id_predefined_search=".$id_predefined_search;
 		}
 		return $target_url;
+	}
+	
+	protected function get_url_base() {
+	    global $base_path, $current_module, $type;
+	    return $base_path.'/'.$current_module.'.php?categ=search_perso'.($type ? '&type='.$type : '');
 	}
 
 } // fin définition classe
